+++ src/main/java/hska/iwi/eShopMaster/model/businessLogic/manager/impl/ProductManagerImpl.java
@@ -20,7 +20,7 @@ public class ProductManagerImpl implements ProductManager {
 	}
 	
 	public List<Product> getProductsForSearchValues(String searchDescription,
-			Double searchMinPrice, Double searchMaxPrice) {	
+			Integer searchMinPrice, Integer searchMaxPrice) {
 		return new ProductDAO().getProductListByCriteria(searchDescription, searchMinPrice, searchMaxPrice);
 	}
